# openapi.yaml
openapi: 3.1.0
info:
  title: LENITNES API
  description: |
    LENITNES is an autonomous AI trading-intelligence operation.
    Frontier-model agents read public commits to consensus-critical
    code, score them against a versioned conviction rubric, execute
    timestamped testnet trades, and broadcast every call to a public
    Telegram channel — all in the same block, all publicly auditable.

    The API exposes:

      • a public surface — scorecard, signals, replay, proof packages,
        public stats, SoSoValue news + macro feeds (no auth required)
      • an operator surface — /admin/* gated by X-Admin-Key
      • a webhook surface — /webhooks/test (HMAC-signed Kraken
        callback was removed in Day 13 when the dead code was
        swept; the autonomous loop drives trades via treasury,
        not via exchange webhooks)
      • infrastructure endpoints — /health, /metrics

    Day 1 pivot (zero-headcount): the per-user surface
    (/auth/*, /rules, /execute/:id) was removed. Every endpoint on
    this spec is post-pivot.
  version: 0.2.0
  contact:
    name: LENITNES
servers:
  - url: https://api.lenitnes.io
    description: Production
  - url: http://localhost:8742
    description: Local development (Docker)
  - url: http://localhost:4000
    description: Local development (npm run dev:api, no Docker)

# No global auth — most endpoints are public. Per-endpoint security
# is declared inline (see `security: []` for public routes and
...
security: []
...
paths:
...
  /admin/status:
    get:
      tags: [Operator]
      summary: Operator status snapshot
      description: |
        Daily signal counts, agent budget spent vs cap, latest signal
        id, treasury wallet count. Returned only when `X-Admin-Key`
        matches `ADMIN_API_KEY` env var. When `ADMIN_API_KEY` is
        unset, returns 503 (no info leak).
      security:
        - adminKey: []
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatus'
        '401':
          description: Invalid admin key
        '503':
          description: ADMIN_API_KEY not configured
  /admin/cache/invalidate:
    post:
      tags: [Operator]
      summary: Invalidate cache entries by pattern
      description: |
        Drops cache entries whose key starts with `pattern`.
        Used when a new signal commits and the scorecard needs to
        refresh before its 60s TTL.
      security:
        - adminKey: []
      parameters:
        - name: pattern
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pattern invalidated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  pattern: { type: string }
                  invalidatedAt: { type: string, format: date-time }
        '400':
          description: pattern query param required
  /admin/cache/invalidate-all:
    post:
      tags: [Operator]
      summary: Invalidate every cache entry
      security:
        - adminKey: []
      responses:
        '200':
          description: All cache entries dropped
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  invalidatedAt: { type: string, format: date-time }
  /admin/risk-check:
    get:
      tags: [Operator]
      summary: Dry-run the trade risk gate without firing a trade
      description: |
        Runs the full risk gate chain (kill switch, asset registry,
        BSC chain-ID guard, balance preflight, position caps, TVL
        floor, volume floor) against the given inputs and returns
        the decision. Used during the first-live-trade preflight
        and ongoing debugging when an expected trade routes to
        paper and the operator wants to know which gate tripped.
      security:
        - adminKey: []
      parameters:
        - name: asset
          in: query
          description: CoinGecko id of the asset (e.g. "bitcoin")
          schema: { type: string }
        - name: chain
          in: query
          description: Chain to evaluate (defaults to bnb)
          schema:
            type: string
            enum: [hedera, arbitrum, robinhood, bnb]
        - name: side
          in: query
          description: Trade side
          schema:
            type: string
            enum: [long, short]
            default: long
        - name: amount
          in: query
          description: Trade size in native chain units (defaults to TREASURY_DEFAULT_AMOUNT)
          schema: { type: string }
      responses:
        '200':
          description: Gate decision (effectiveMode + reason)
          content:
            application/json:
              schema:
                type: object
                properties:
                  input:
                    type: object
                    properties:
                      asset: { type: string }
                      chain: { type: string }
                      side: { type: string }
                      amountIn: { type: string }
                      intendedMode: { type: string, enum: [paper, live] }
                  decision:
                    type: object
                    properties:
                      effectiveMode: { type: string, enum: [paper, live] }
                      reason: { type: string }
                      downgraded: { type: boolean }
                  tradingEnabled: { type: boolean }
        '401':
          description: Invalid admin key
        '503':
          description: Admin not configured
  /admin/positions/{id}/close:
    post:
      tags: [Operator]
      summary: Manually close an open position
      description: |
        Fetches the current CoinGecko price for the position's
        asset, then closes the position. Fires a real on-chain
        reverse swap when TRADING_ENABLED + the asset is in the
        registry + the open was on-chain; otherwise records a
        paper close. Returns realized PnL + close tx hash.
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          description: Position UUID
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Close result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  positionId: { type: string, format: uuid }
                  asset: { type: string }
                  exitPriceUsd: { type: number, nullable: true }
                  pnlUsd: { type: number, nullable: true }
                  closeTxHash:
                    { type: string, nullable: true, description: 'Null for paper closes' }
        '404':
          description: Position not found
        '409':
          description: Position is not open
        '500':
          description: Close failed (swap reverted or DB error)

  # ───────────────────────────────────────────────────────────
  # SoSoValue on-chain finance data
  # ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Matches the `ADMIN_API_KEY` env var. When unset, all
        /admin/* endpoints return 503 (no info leak).
  schemas:
...
    Chain:
      type: string
      enum: [hedera, arbitrum, robinhood, bnb, valuechain]
...
    AdminStatus:
      type: object
      required: [signals, agent, trades, treasury]
      properties:
        signals:
          type: object
          required: [last24h, last7d]
          properties:
            last24h: { type: integer }
            last7d: { type: integer }
            latestAt: { type: string, format: date-time, nullable: true }
            latestId: { type: string, format: uuid, nullable: true }
        agent:
          type: object
          required: [scoresLast24h, dailySpendUsd, dailyBudgetUsd]
          properties:
            scoresLast24h: { type: integer }
            dailySpendUsd: { type: number }
            dailyBudgetUsd: { type: number }
        trades:
          type: object
          required: [filledAllTime]
          properties:
            filledAllTime: { type: integer }
        treasury:
          type: object
          required: [activeWallets, defaultChain, defaultMode]
          properties:
            activeWallets: { type: integer }
            defaultChain: { $ref: '#/components/schemas/Chain' }
            defaultMode:
              type: string
              enum: [paper, live]

    # ── DLQ ─────────────────────────────────────────────────
...